txc_dxtn--, LuaJIT--, OZ_NET--, ozCore misc++
- removed dummy txc_dxtn, "force_s3tc_enable=true" does the trick
- removed LuaJIT support, it cannot execute precompiled code
- commented out network code and removed OZ_NET
- updated README
- ozCore
  * mCopy, mMove & co. accept int instead of size_t
  * containers: some fixes of resize(), element destruction (aFill() is not OK)

+++ src/client/Model.cc
@@ -476,7 +476,7 @@ const File* Model::preload()
         normals[i] = is.readVec3();
       }
 
-      mCopy( vertices, vertexBuffer, size_t( nVertices ) * sizeof( Vertex ) );
+      mCopy( vertices, vertexBuffer, nVertices * int( sizeof( Vertex ) ) );
     }
   }
