+++ src/tests/noise.cc
@@ -76,8 +76,8 @@ int main( int, char** )
   Log() << "b2.length() = " << b2.length();
   Log() << "b3.length() = " << b3.length();
 
-  Log() << "b0 == b2: " << ( mCompare( b0.begin(), b2.begin(), size_t( b0.length() ) ) == 0 );
-  Log() << "b1 == b3: " << ( mCompare( b1.begin(), b3.begin(), size_t( b1.length() ) ) == 0 );
+  Log() << "b0 == b2: " << ( mCompare( b0.begin(), b2.begin(), b0.length() ) == 0 );
+  Log() << "b1 == b3: " << ( mCompare( b1.begin(), b3.begin(), b1.length() ) == 0 );
 
   bool      isAlive = true;
   SDL_Event event;
